requirements.txt
pandas
numpy
sqlalchemy
psycopg2-binary
requests
dotenv
logging